HaJongsu/requirements.txt
# Qook 신선식품 챗봇 의존성 패키지

# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# LangGraph 및 LLM
langgraph==0.0.64
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.6.1

# 데이터베이스
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1

# 벡터 스토어
chromadb==0.4.18
faiss-cpu==1.7.4

# 외부 API
tavily-python==0.3.3
requests==2.31.0

# 데이터 처리
pandas==2.1.4
numpy==1.26.2

# 환경 변수
python-dotenv==1.0.0

# 로깅 및 모니터링
structlog==23.2.0

# 개발 도구
pytest==7.4.3
black==23.12.0
flake8==6.1.0

# 보안
cryptography==41.0.8

# 기타 유틸리티
pydantic==2.5.2
python-dateutil==2.8.2
pillow==10.1.0